fix conditional rendering problem
